Struct pi_lcd::HD44780

pub struct HD44780 {
    // some fields omitted
}

The display handle

Methods

impl HD44780

fn new(disp_rs: u32, disp_e: u32, datalines: [u32; 4], disp_cols: u32, disp_rows: u32) -> HD44780

Creates a new HD44780 instance with disp_rs as rs pin, disp_e as enabled pin, datalines as data4 to data7

disp_cols are the number of columns disp_rows are the number of rows

fn init(&self)

Initializes the display and clears it

fn clear(&self)

Clears the display

fn command(&self, bits: u8)

Sends a given byte as a command

fn send_string(&self, text: String, row: u32)

Parses a String and and outputs it to the given row

fn create_char(&self, address: u8, bitmap: [u8; 8]) -> Result<u8, &'static str>

Creates a new custom character from a bitmap on the given address